"models setting, exceptions, bugs"

+++ BooksS/src/main/java/com/library/booksS/controllers/AccountController.java
@@ -2,9 +2,9 @@ package com.library.booksS.controllers;
 
 import com.library.booksS.repositories.AccountRepository;
 import com.library.booksS.models.AccountB;
+import com.library.booksS.exceptions.AccountBAlreadyExistsException;
 import com.library.booksS.exceptions.AccountBNotFoundException;
 import org.springframework.web.bind.annotation.*;
-
 
 @RestController
 public class AccountController {
@@ -16,7 +16,12 @@ public class AccountController {
 
     @PostMapping("/Account")
     AccountB newAccount(@RequestBody AccountB account) {
-        return accountRepository.save(account);
+        AccountB accountc = accountRepository.findById(account.getIdUser()).orElse(null);
+        if(accountc != null){
+            throw new AccountBAlreadyExistsException("Account already exists");
+        }else{
+            return accountRepository.save(account);
+       }
     }
 
     @GetMapping("/Account/{idUser}")
@@ -24,20 +29,30 @@ public class AccountController {
         AccountB account = accountRepository.findById(idUser).orElse(null);
 
         if (account == null) {
-            throw new AccountBNotFoundException("No se encontro una cuenta con el id: " + idUser);
+            throw new AccountBNotFoundException("User not found");
         }
 
         return account;
     }
 
-    @PostMapping("/Account/update")
-    AccountB updateAccount(@RequestBody AccountB account) {
-        return accountRepository.save(account);
+    @PostMapping("/Account/update/{idUser}")
+    AccountB updateAccount(@RequestBody AccountB account, @PathVariable int idUser) {
+        
+        if(account.getIdUser() != idUser){
+            throw new AccountBNotFoundException("Invalid credentials");
+        }else{
+            return accountRepository.save(account);
+        }
+
     }
 
-    @DeleteMapping("/Account/delete")
-    void deleteAccount(@RequestBody AccountB account) {
-        accountRepository.delete(account);
-        return;
+    @DeleteMapping("/Account/delete/{idUser}")
+    void deleteAccount(@RequestBody AccountB account, @PathVariable int idUser) {
+        
+        if(account.getIdUser() != idUser){
+            throw new AccountBNotFoundException("Invalid credentials");
+        }else{
+            accountRepository.delete(account);            
+        }
     }
 }
