+++ Rental_ms/src/main/java/com/LibreriaRentalApp/Rental_ms/controllers/AccountController.java
@@ -2,6 +2,7 @@ package com.LibreriaRentalApp.Rental_ms.controllers;
 
 import com.LibreriaRentalApp.Rental_ms.repositories.AccountRepository;
 import com.LibreriaRentalApp.Rental_ms.models.AccountB;
+import com.LibreriaRentalApp.Rental_ms.exceptions.AccountBAlreadyExistsException;
 import com.LibreriaRentalApp.Rental_ms.exceptions.AccountBNotFoundException;
 import org.springframework.web.bind.annotation.*;
 
@@ -15,7 +16,12 @@ public class AccountController {
 
     @PostMapping("/Account")
     AccountB newAccount(@RequestBody AccountB account) {
-        return accountRepository.save(account);
+        AccountB accountc = accountRepository.findById(account.getIdUser()).orElse(null);
+        if(accountc != null){
+            throw new AccountBAlreadyExistsException("Account already exists");
+        }else{
+            return accountRepository.save(account);
+       }
     }
 
     @GetMapping("/Account/{idUser}")
@@ -23,21 +29,29 @@ public class AccountController {
         AccountB account = accountRepository.findById(idUser).orElse(null);
 
         if (account == null) {
-            throw new AccountBNotFoundException("No se encontro una cuenta con el id: " + idUser);
+            throw new AccountBNotFoundException("User not found");
         }
 
         return account;
     }
 
-    @PostMapping("/Account/update")
-    AccountB updateAccount(@RequestBody AccountB account) {
-        return accountRepository.save(account);
+    @PostMapping("/Account/update/{idUser}")
+    AccountB updateAccount(@RequestBody AccountB account, @PathVariable int idUser) {
+        if(account.getIdUser() != idUser){
+            throw new AccountBNotFoundException("Invalid credentials");
+        }else{
+            return accountRepository.save(account);
+        }
     }
 
-    @DeleteMapping("/Account/delete")
-    void deleteAccount(@RequestBody AccountB account) {
-        accountRepository.delete(account);
-        return;
+    @DeleteMapping("/Account/delete/{idUser}")
+    void deleteAccount(@RequestBody AccountB account, @PathVariable int idUser) {
+
+        if(account.getIdUser() != idUser){
+            throw new AccountBNotFoundException("Invalid credentials");
+        }else{
+            accountRepository.delete(account);            
+        }
     }
 
 }
